\name{as.spherical}
\alias{as.spherical}
\title{Rescale matrix rows to have unit length}
\description{
  The function rescales each row of the matrix to have unit length
}
\usage{
  as.spherical(X)
}
\arguments{
  \item{X}{matrix}
}
\value{
  a spherical matrix with the same dimension as the input matrix
}
